refactor(editor): Move the RAG template opening helper to reusable composable
